rename march module to uarch
